pyproject.toml
[project]
name = "cbi-policy-compliance-rag"
version = "0.1.0"
description = "Arabic-first OCR + RAG compliance checker for Central Bank of Iraq policy documents"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "Open Source Maintainer"}]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "sqlalchemy>=2.0.31",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.2",
  "openai>=1.40.0",
  "faiss-cpu>=1.8.0.post1",
  "numpy>=1.26.4",
  "paddleocr>=2.7.3",
  "pytesseract>=0.3.10",
  "opencv-python-headless>=4.10.0.84",
  "pymupdf>=1.24.9",
  "pillow>=10.4.0",
  "orjson>=3.10.6",
  "httpx>=0.27.0",
  "python-json-logger>=2.0.7"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.5",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"
